+++ nightfall-client/src/classes/index.mjs
@@ -2,6 +2,5 @@ import Transaction from 'common-files/classes/transaction.mjs';
 
 import Commitment from './commitment.mjs';
 import Nullifier from './nullifier.mjs';
-import Secrets from './secrets.mjs';
 
-export { Transaction, Commitment, Nullifier, Secrets };
+export { Transaction, Commitment, Nullifier };
