+++ nightfall-client/src/services/commitment-sync.mjs
@@ -6,12 +6,12 @@ or use clientCommitmentSync to decrypt when new zkpPrivateKey is received.
 import config from 'config';
 import logger from 'common-files/utils/logger.mjs';
 import { generalise } from 'general-number';
+import { edwardsDecompress } from 'common-files/utils/curve-maths/curves.mjs';
 import { getAllTransactions } from './database.mjs';
 import { countCommitments, storeCommitment } from './commitment-storage.mjs';
 import { decrypt, packSecrets } from './kem-dem.mjs';
 import { ZkpKeys } from './keys.mjs';
 import Commitment from '../classes/commitment.mjs';
-import { edwardsDecompress } from '../utils/crypto/encryption/elgamal.mjs';
 
 const { ZERO } = config;
